again_zappa/requirements.txt
aniso8601==9.0.1
argcomplete==1.9.2
attrs==22.1.0
base58==0.2.4
boto3==1.23.10
botocore==1.26.10
cachetools==4.2.4
certifi==2022.9.24
cfn-flip==1.3.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.5
dataclasses==0.8
docutils==0.18.1
durationpy==0.5
Flask==2.0.3
flask-restplus==0.13.0
future==0.16.0
google-api-core==2.8.2
google-auth==2.14.1
google-cloud-dialogflow==2.14.1
googleapis-common-protos==1.56.3
grpcio==1.48.2
grpcio-status==1.50.0
hjson==3.0.1
idna==3.4
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.9.3
jsonschema==4.0.0
kappa==0.6.0
lambda-packages==0.19.0
MarkupSafe==2.0.1
placebo==0.10.0
proto-plus==1.22.1
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyrsistent==0.18.0
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2022.6
PyYAML==3.12
requests==2.27.1
rsa==4.9
s3transfer==0.5.2
six==1.16.0
toml==0.10.2
tqdm==4.19.1
troposphere==3.1.1
typing-extensions==4.1.1
Unidecode==1.3.6
urllib3==1.26.12
Werkzeug==2.0.3
wsgi-request-logger==0.4.6
zappa==0.45.1
zipp==3.6.0
